Sessionize data updates
